Replaced shift_filter_func with shift_threshold param
Most cases you just want to change the threshold and not the function. If you want to look for negative or positive shifts you can now pass shift_sign as well. This change simplifies the coming optimisation pipeline. If users need more advanced filtering because they are using some special shift detection algorithm, this can still be done manually, or using the new preprocess_variable() which allows for easily applying a lambda function to a certain variable.