feat: cftime support
cftime is a very common time format for CMIP datasets. Shifts and clustering computations however need int/float time values. So does stat functions. To avoid converting the time values the user's dataset, we add a new property to the toad object called numeric_time_values, which, if the dataset has cftime time values, will be converted to seconds since the start of the simulation.

Keeping the cftime format is also useful for plotting.

* New functions in core.py:
- _is_time_numeric()
- numeric_time_values
- numeric_time_values_unit
* get_cluster_mask() and apply_cluster_mask() now allows you to use numeric_times.